+++ Samples_OpenNI_Cpp/TrackHand/src/main.cpp
@@ -151,8 +151,8 @@ int main(int argc, char* argv[]) {
 
     if (hand_recognized) {
       // Draw point over tracked hand
-      glue.DrawPointOverRegion(projective_point.X, projective_point.Y, 0, 0);
-      glue.DrawPointOverRegion(projective_point.X, projective_point.Y, 320, 0);
+      glue.DrawPointOverRegion(static_cast<unsigned>(projective_point.X), static_cast<unsigned>(projective_point.Y), 0, 0);
+      glue.DrawPointOverRegion(static_cast<unsigned>(projective_point.X), static_cast<unsigned>(projective_point.Y), 320, 0);
     }
 
     glue.EndDraw();
